backend/requirements.txt
fastapi
uvicorn
python-multipart
python-multipart
pillow
pillow-heif
